MB-56181: Improve the reliability of forceMaxCas
The forceMaxCas path exists to allow the repair of a hybrid logical
clock that has been exposed to a bad (future) timestamp. In the
previous implementation only the in-memory max_cas (part of the HLC
object) is fixed, this left the disk max_cas as the bad value. A
warm-up would then undo the fix and put the user back to the original
bad state.

Two changes occur in this fix.

1) forceMaxCas triggers a new set-vb-state meta mutation to be queued
in the checkpoint. This is the start of the process of getting the
disk max_cas repaired.

Even with this new meta-mutation bringing the repaired cas down to
the flusher - it was noted that the fix was ignored if data was also
flushing. The flusher itself would only store the max value it observes
i.e. the bad value (or at least cas values based on the bad HLC) is
just stored again. The fact that the flush batch is reordered by key
is also troublesome for a fix.

2) The tracking of the max_cas seen in a flush batch is now moved out
of the main flush loop. This is because the flush loop is processing
items in key, not seqno order. The max_cas is now processed when the
checkpoint items are visited (which is in seqno order) and this allows
the forceMaxCas path to inject a fixed CAS, push a set-vbucket-state
and reset the max_cas for flushing to the new CAS.

Change-Id: Ic7d0aba8d9e62b76623af55ddddb5baf9b6aec10
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/189025
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>