Make the quota-sharing tests less flaky
These tests rely on being able to increase mem_used to a specific value
and trigger paging deterministically.

Because mem_used also counts the CheckpointManager, intermittently, CM
memory may cause eviction to run earlier or for us to evict more items.

Configure the test buckets such that we immediately try to expel items
and destroy checkpoints, and wait for this to happen before returning
from generateItems to make the tests less flaky.

Also, set watermarks after we've populated the bucket with items, as
this is less likely to trigger paging due to mem_used spikes.

Change-Id: I6864228d6e18cdaf55f61a0b65c4a9119d9f7ded
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/189024
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>