MB-56256: CDC don't deduplicate abort/prepare
When flushing a historical batch the flusher must not deduplicate
an abort/prepare (for collections that allow deduplication).

Maintaining the sequence of abort(k1), prepare(k1) is critical for the
consistency of the replica, where it does not expect
prepare(k1), prepare(k1) without an interleaving abort or commit.

Change-Id: I16cd667cc26047aecea2ccf60c6516963dc210ba
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/189083
Reviewed-by: Dave Rigby <daver@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Jim Walker <jim@couchbase.com>