MB-56256: Don't deduplicate away abort/commit for historical flush
The sequence index must ensure that every prepare in the history
window has at least one resolution, that is either a commit or
an abort.

This commit is possibly over-cautious, there maybe some deduplication
which can occur provided that never is there a prepare->prepare
sequence in the history window.

For example a sequence of operations which broke prior to this fix is.

* 1: prepare(k)
* flush
* 2:commit(k)
* 3:prepare(k)
* 4:commit(k)
* flush

In the second flush batch 2:commit(k) is deduplicated, leaving the
following sequence in history.

* 1:prepare(k)
* 3:prepare(k)
* 4:commit(k)

In terms of over-cautious, the following sequence could have some level
of deduplication still applied.

* 1:prepare(k)
* 2:commit(k)
* 3:prepare(k)
* 4:commit(k)
* flush

Leaves the 4 items in the history index, whereas it could be reduced to
just two items.

* 3:prepare(k)
* 4:commit(k)

But instead 3 items will remain.

* 2:commit(k)
* 3:prepare(k)
* 4:commit(k)

Change-Id: I5c55ed0de4ebce867ccbc2a2ba49504c3fbdef0b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/189152
Well-Formed: Restriction Checker
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>