MB-56275: DcpStreamEnd must ensure buffered bytes are acknowledged
If a DcpStreamEnd occurs and the PassiveStream is currently buffering
incoming data (awaiting memory reduction), prior to this fix a
DcpStreamEnd message would remove and destruct the PassiveStream but
not acknowledge back to the producer the flow-control bytes.

With this commit the DcpStreamEnd handler will move any outstanding
value out of the PassiveStream and ensure they are accounted for.

Change-Id: I41d63e56461de73e3221aabfec5e4e36156b399d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/189415
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>