MB-55396: Tag metrics with version added
Version determined programatically by comparison with historical
stats.def.h files; the first release version a given key appears in has
been taken as the release version here.

Going forward, this should be included in any newly added metric.
This is not yet enforced, but will be in a later patch.

Change-Id: Ic3468e1be8369222b0e8d2bc233e8fd16135ef83
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/189497
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: James H <james.harrison@couchbase.com>