MB-55396: Tag metrics with stability
To support metrics documentation, a "stability" level has been requested
for each metric.

This may be {"committed","volatile","internal"}.

Only committed metrics are expected to be deprecated if removed or
changed.

If missing, this will be considered to default to "committed".

This patch tags elixir and rocksdb metrics as "volatile"/"internal".

Change-Id: I1b625fb9b801ae59f2c48b4f0619af449b70af45
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/191514
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: James H <james.harrison@couchbase.com>