MB-56148: Fix the DcpProducer notify logic to consider seqno-advance
Prior to this commit a DcpProducer could not receive a wakeup for
prepare/abort when it should to trigger a seqno advance message.

The notifySeqnoAvailable early return is now updated to check if
the producer is capable of seqno-advance. For example with this commit
a producer created by GSI, which supports seqno-advance but not sync
writes will be awoken for

* normal mutations/commits
* aborts

But never prepares (which will become commit or abort).

Change-Id: I217ca7a934217a325b3c0c918fb3817f6db167ca
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/189535
Well-Formed: Restriction Checker
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>