MB-55504 Fix per doc compression
Per doc compression was broken for following reasons:
* In case of memory optimised writes, KV would account for the
  compressed size in its disk size stat, whereas Magma would persist the
  uncompressed value to sstable.
* In case of write batches carrying duplicate keys, the deltas we
  generate would carry uncompressed value sizes, even though they would
  undergo compression.

The linked Magma change fixes these two issues:
https://review.couchbase.org/c/magma/+/189637

Magma now calls DocTransformCallback before the WriteDocsCallback. Hence
the disk size stat updates can be done entirely inside WriteDocsCallback
as-is. This simplifies KV's DocTransformCallback implementation and
makes the maybeCompressValue idempotent, which is now a requirement as
Magma invokes the callback on the same document twice in case of memory
optimised writes.

Perf testing shows there is no regression when we get a good compression
ratio as the time spent in compression is offset by lesser amount of
data we have to write to WAL.

Change-Id: I290e59c339b3d9c6d196c9c432fab405086c3081
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/189645
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>