MB-56452: CDC backfill must not register a cursor twice
When CDC backfill consists of data from two scans and includes
two snapshot markers, the second scan/markDiskSnapshot must not
re-register the checkpoint cursor.

If it does re-register the cursor (same seqno) it risks triggering
eager checkpoint removal, which in turn will mean the second
register call "fails", losing the memory stream and requiring a
second backfill.

This is fixed by indicating to markDiskSnapshot if the history
snapshot is a single snapshot or snapshot combining
non-history/history.

Change-Id: I9277fe62c8119c25e46a733365d874e7c0741d3f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/189675
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>
Well-Formed: Restriction Checker