MB-56547 [1/3]: Add option to disable DCP OSO backfill at runtime
As per analysis in MB-56346, out-of-seqno (OSO) backfill is not always
beneficial compared to seqno backfill - for larger collections and/or
small value sizes it can be significantly slower than "classic" seqno
backfill.

We need to modify the logic of when OSO backfill is performed to not
only consider when it is possible , but also when it is expected to be
faster than seqno.

As an initial step for that, add a new config param
(ep_dcp_oso_backfill=enabled|disabled) which controls if OSO should be
used for backfills or not. This can be changed dynamically via epctl.

Change-Id: I8d558a2ac381ae5070d36bf185c6a6611abd3a62
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/189920
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>