MB-56565: Defer removal of checkpoint until after cursor creation
If DCP is cursor dropped it entered a perpetual cycle of backfilling
as eager checkpoint removal was triggered when markDiskSnapshot
attempts to register a cursor. This loop occurs when the checkpoint
that DCP needs is unreferenced at the time, and it inadvertently
triggers its removal just before it tries to reference it.

This fix ensures that in the registerCursorBySeqno path that checkpoint
removal is triggered after the cursor is created and referencing a
checkpoint, ensuring it doesn't remove the required checkpoint.

Change-Id: Id408b40bff5b6b079893ee80d4c621aa28dac86c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/189976
Reviewed-by: James H <james.harrison@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Jim Walker <jim@couchbase.com>