MB-56094: Add 'cursor_op' to CM cursors stats
We already dump the seqno, the 'operation' is another useful
information at diagnostic. In particular in the case where cursor points
to a meta-item, as consecutive meta-items share the same seqno.

Output example:

  vb_0:persistence:cursor_op:             set_vbucket_state
  vb_0:persistence:cursor_seqno:          1

Note: In the example above the checkpoint is like [e:1 cs:1 vbs:1), so
the seqno isn't enough for knowing the precise cursor position.

Change-Id: I280eb9789eb45e2ecdfb51df0fbabe75801655aa
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/189981
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>