Use correct numeric types for the ArrayHistogram calculations
`float` only has 23-bit long mantissa, so the maximum integer that can
be safely represented is 8 388 608, which is too low for correct
calculation of MFU thresholds (we can easily have more than 8 million
items in vBuckets).

`std::aggregate` uses the type of the last parameter, which was `int` in
this case. Not sufficient to hold the sum of all bucket counts given a
generic `CountType`.

Change-Id: I93f9af90ecc420af5355c6f82eda9a8c837e5d90
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/190045
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>