MB-56672: Limit concurrency on various tasks
Currently the memory cost for each hash specified for each argon2id
hash is set to 8MB, but it may be overridden via the
/settings/security/argon2idMem endpoint. We want to limit the number
of concurrent hashes (from a CPU and memory perspective) and this
patch introduce a limit of max 6 concurrent authentication tasks.

Introduced a new config key:

   "max_concurrent_authentications" : <number>

in memcached.json to let ns_server tune the number when the memory
cost change.

The number 6 was picked as the authentication runs in the NonIO
executor pool which may grow up to 8 threads. In a worst case
scenario one would use up to 48MB of memory without any tuning.
It would be up to ns_server to "correctly" adjust this number
based off the max value it used for generating the hash values.

Change-Id: I44d5f94d5b3212608354d07d47c8356f6e96a131
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/190183
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>