MB-56094: Handle correctly unref checkpoints at CM::registerCursor
If the user tries to register a cursor that already exists (by name),
then the existing cursor is removed from CM. That step might make some
checkpoint eligible for removal and deallocation.

Before this fix, we miss to handle those removed checkpoints properly,
ie we ingore the return value of the CM::removeCursor call and we miss
to pass them to the CheckpointDestroyer. The effects of that are:

- First and most important, we miss to update some mem stats in CM that
  are used for enforcing the CheckpointQuota (main issue solved here,
  core of MB-56094)
- Secondly, checkpoints are released in-place under CM-lock, which is
  suboptimal as checkpoint deallocation is a slow operation

Change-Id: I47a0f56f4d96453cc9b02cf1d9cf9996e43aceff
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/190352
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>