Use relaxed atomic operations for connection stats
We do not require anything more than relaxed semantics for stats.

RelAcq semantics are most costly on Arm (as they use load/store
linked and require extra synchronization with preceding operations).

Also upgraded consume -> acquire in settings, as use of consume is
discouraged and it doesn't provide the ordering guarantees we need.

Change-Id: Icdfda6bdf156999cb76bc76920b271b946e00be3
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/190917
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>