MB-55357: Validate Marker's visible-snap-end at PassiveStream
If replica receives snapEnd<visibleSnapEnd in a SnapshotMarker, then we
intercept that invalid state at checkpoint creation when we have
already closed the open checkpoint but before we create the new open
one.
That leaves the CheckpointManager in an invalid state where there's no
open checkpoint. A subsequent call into the CM (into any path that
validates the CheckpointList) might crash the node at that point.
That's the exact cause of the crash in MB-55357.

This change adds validation at PassiveStream level. That doesn't solve
the root issue (ie, the fact that a Producer made/sent an invalid
Marker to a replica node), but that prevents a crash as the validation
is perfomed before manipulating the CheckpointList, so only the
affected DCP connection is closed and the node stays up.

Change-Id: Ib24efee180808df9d2378dea7a766ea2b49b6d8b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/190913
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>