MB-56893: Unauthenticated users should have a max size of 1k
We don't want unauthenticated clients to start spooling large
packets on the server. 1k was chosen as that is the minimum chunk
size in libevent when built as LP64. HELO and SASL auth messages
do not exceed 1k.

This patch only address the libevent backend (as that's the one
used in production and the folly backend is still experimental)

Change-Id: Ia4e1bfe6ad377a4bc7a07a57afc01aeb44b5fd2c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/190982
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>