MB-56891: DcpConnMap::each(): unlock map before calling func
As observed in MB-56770 - during a rebalance when a client (query) was
disconnecting and reconnecting to KV-Engine ~30 times per second,
front-end threads which are trying to remove a closed connection can
be blocked by STAT commands running on other connections -
specifically dcp and dcpagg stat commands.

This is because both dcp and dcpagg call DcpConnMap::each to print
stats for every connection, and DcpConnMap::each currently takes a
write (exclusive) lock on the cookieToConnMap for the duration of the
call.

We can reduce this cost by not locking the cookieToConnMap for the
entire call; instead just copy (a reference to each ConnHandler) in
cookieToConnMap into a temp container, then unlock the cookieToConnMap
before actually invoking the functions.

Change-Id: I6c0811aa53149949f4255dece95c9a710a4f8235
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/191090
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>