MB-56891: DcpConnMap::each(): use shared lock during copy
Modify DcpConnMap::each() to take a copy of the underlying set of
connections using a read (shared) lock instead of an exclusive one.

This allows multiple callers (e.g. multiple STAT commands executing
concurrently) to all take a copy of the map of connections
concurrently, without one blocking the other.

This is likely to be of only marginal benefit as the previously
exclusive lock duration _should_ be short; however it is O(n) with the
number of DCP connections so for large deployments the exclusive lock
period might start to become significant.

Note: this doesn't change the semantics of the each() iteration - the
      callback function is still invoked on a non-const instance of
      the element (std::shared_ptr<ConnHandler>). There's potentially
      further improvements in terms of not passing around the owning
      std::shared_ptr<> elements to the passed in function (which
      could reduce refcount incr/decr), and adding a const overload of
      each() which in turn would allow const versions of methods used
      by users of each() to be called.

Change-Id: I1b69c2af69ca25853a61df94009b8ee73e35af1d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/192449
Well-Formed: Restriction Checker
Reviewed-by: James H <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>