MB-56855: RangeScan continue is ending too early
Changes made for MB-55672 coded for a case where the I/O task uses
success as the  I/O complete status - which would indicate that the
task reached the send "buffering" limit, but the continue itself is not
complete. However code was missing to actually push that status via
notifyIOComplete meaning that when the scan yielded for the buffering
limit, the entire continue request ended with a "more" status,
meaning the client must issue another continue.

This fix is made in relation to MB-56855 as this bug was noted because
in theory if the blocked worker thread was awoken with a
"Cookie::getAioStat" value of success it would lead to MB-56855, a
worker thread which doesn't get an I/O complete notification eventually
blocking bucket delete. This occurs because the frontend code in
RangeScanOwner forgets to re-copy the Cookie* into the RangeScan when
scheduling for the success case.

This commit also adds a new Expect into ~RangeScan to help detect if
MB-56855 re-occurs.

Change-Id: Ic05d2e078984427fd4333fcfad3381ffbc0bf76e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/191024
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>