MB-56855: Improve RangeScan bucket delete handling
The changes in this commit were suspected to be required to address
MB-56855, but the new unit-tests suggests these changes are not
required. However the unit-test and code change are useful
improvements.

Both RangeScan I/O tasks should not set "blockShutdown=false", they
should both block shutdown to ensure clients are notified of IO
complete.

Change-Id: I67da6faf0335f389fee2aa7bbcefad3c3ba87685
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/191029
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>