MB-56970: Ensure deleteWithMeta does not bgfetch for non-existent items
As a result of MB-25671/MB-36087, deleteWithMeta would try to bgfetch a
value for any non-resident item with datatype==xattrs.

TempNonExistent items indicate that a bgfetch did not find a value on
disk. As there is no value, the datatype is largely irrelevant.

However, the datatype is not cleared/reset when transitioning
tempDeleted items to tempNonExistent. This allows a tempNonExistent item
with datatype==xattrs to be created.

Upon encountering such an item, deleteWithMeta would enter an endless
bgfetch loop - the bgfetch would not find a value on disk, but once
notified the frontend would again find the tempNonExistent item, and
again request a bgfetch.

As a minimal fix for this specific case, handle tempNonExistent items
explicitly in deleteWithMeta.

A wider review of temp item handling will follow, as will investigation
of limiting ewouldblock cycles, and a refactor of deleteWithMeta.

Change-Id: I4e96472bbc198cfedb40c33d7e934d47433d0f0e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/191280
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: James H <james.harrison@couchbase.com>