MB-56855: Don't allow RangeScanCancel to clear the cookie
MB-56855 occurs because RangeScanCancel can run concurrently with
a RangeScanContinue, the continue should be on its way to detecting
the cancel state, but nothing stops a cancel from proceeding. If this
does occur, the RangeScanCancel will call prepareToContinueOnIOThread
an detect the cancel, but it's too late the continueRunState is over
written by the cancel as the first line of the function, wiping out
continueRunState.cState.cookie. Now when the continue task returns
it will skip notifying the client.

The main fix here is to only update continueRunState when the task
is on the continue path, now a concurrent cancel will not overwrite
the cookie. The fix also includes simplification of the
ContinueRunState class, it does not need to keep the cookie in the
object, instead the cookie is returned to via
prepareToContinueOnIOThread so it sits safely on the stack of the
task running the continue.

Change-Id: I8b13ad3046296fa2342044a3564b0b39af8d8b25
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/191299
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>