MB-57002: Disable TLS session resume (NO_TICKET)
If a SDK client connection attempts to negotiate a TLS connection, and
the client supports TLS Session Resume (RFC 5077) client certs, and
client certs are enabled (Security -> "Require Client Certificate" ->
Enabled), then KV-Engine fails to establish the connection and client
bootstrap fails with the following OpenSSL error:

    WARNING 68: Unrecoverable error encountered: ["reading","error"], ssl_error: [{error:00000001:lib(0):func(0):reason(1)},{error:140D9115:SSL routines:ssl_get_prev_session:session id context uninitialized}], shutting down connection

Such Couchbase clients include .NET 6 applications running on Windows
10 Pro, as that automatically enables TLS Session Resume.

The problem is that OpenSSL does not fully enable TLS Session Resume
by default - it happily issues session tickets if the client asks for
them, but it does not specify a default session_id, which effectively
means the session ticket cannot be used for resumption.

Normally this is benign (the connection is established regardless),
but _if_ client certs are enabled (set SSL_VERIFY_PEER option), then
OpenSSL returns a fatal error if the client attempted to use a session
ticket (as shown above):

    int ssl_get_prev_session(SSL_CONNECTION *s, CLIENTHELLO_MSG *hello)
    {
    ...
        if ((s->verify_mode & SSL_VERIFY_PEER) && s->sid_ctx_length == 0) {
            /*
             * We can't be sure if this session is being used out of context,
             * which is especially important for SSL_VERIFY_PEER. The application
             * should have used SSL[_CTX]_set_session_id_context. For this error
             * case, we generate an error instead of treating the event like a
             * cache miss (otherwise it would be easy for applications to
             * effectively disable the session cache by accident without anyone
             * noticing).
             */

            SSLfatal(s, SSL_AD_INTERNAL_ERROR,
                     SSL_R_SESSION_ID_CONTEXT_UNINITIALIZED);
            fatal = 1;
            goto err;
        }

There's essentially two ways to fix this issue:

1) Fully setup TLS session resume in KV-Engine.
2) Disable issueing of TLS session tickets.

While TLS session resume is in theory a useful feature - it reduces
the number of network round-trips needed to setup a subsequent
connection if the client has already connected to the same server - it
also carries some potential risks (particulary in TLSv1.2 and lower)
as it reduces the forward security of a connection (session keys are
re-used across multiple connections).

Also, our mcbp connections are normally long-lived and so the benefit
of session resume is mimimal.

It also requires code in each SDK client to enable it (other than .NET
using Windows Schannel, which _does_ enable it by default).

As such, this patch simply disables it for now - it could only be used
by .NET SDK clients on Windows right now anyway, where it was broken
due to our missing setup.

Note I haven't included a test as part of this patch. I did
investigate it, but it would require significant work - I would
essentially have to expand our mcbp TLS connection to implement full
support session resume, attempt it against KV-Engine and then fail the
test if resume succeeded. That seemed excessive given we are simply
using a standard OpenSSL flag to turn a feature off.

Change-Id: I23f6218ec488e06aef6cf31f66f7f185de1a5c85
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/191354
Well-Formed: Restriction Checker
Reviewed-by: James H <james.harrison@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>