MB-57064: Revert "MB-47952: Don't expire temp items"
This reverts commit b8ca1ccf4bc1ae92c48940eda0376582b805569a.

Reason for revert: In value-only eviction, we cannot remove
tempDeleted / tempNonExistent items - they will eventually grow to
consume the entire bucket quota.

Prior to MB-47952, this was not an issue as temp items were cleaned up
("expired") via the ExpiredItemPager, however that was removed in
MB-47952 so temp items were not accounted as expiries. The expectation
was that the ItemPager would handle the temp item cleanup, however
this was incorrect for value-eviction - see
StoredValue::eligibleForEviction():

    bool eligibleForEviction(EvictionPolicy policy) const {
        ...
        if (policy == EvictionPolicy::Value) {
            return isResident() && !isDirty();
        } else {
            return !isDirty();
        }
    }

Given that temp items are not considered resident, we always return
false in value-eviction - and hence there is no way to remove temp
items.

For MP, simply revert the change to no expire temp items, to restore
the pre-7.1.0 behaviour.

Change-Id: Ie45a32e28789b4fba518bfe2cb6f4628616f5acb
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/190813
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>