MB-56934: Add "kv_" prefixed metering metrics in _prometheusMetrics
It was decided that metering metrics should not be service-prefixed (as
all other metrics are).

This remains the case for metrics in the _metering endpoint, but for the
copies in _prometheusMetrics (for local Prometheus consumption), it has
been decided that they should indeed be consistent with all other kv_
prefixed metrics.

For now, expose two copies of the _metering metrics. There are not many,
and the unprefixed versions will be removed in MB-56935, after consumers
have moved to the updated names.

Change-Id: I2b32beb12af98539081fc4145445fff1741895de
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/191445
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>