Do not increment per-vBucket getOps stat when queueing ops
We return EWB when the vBucket is pending and retry the op. The number
of pending ops is counted separately.

Additionally, this behaviour differs from what we do for the
per-collection counters, which we increment when status != would_block,
so this behaviour can introduce skew.

We don't need to increment this stat here (and it's
inconsistent/incorrect to do so). Removing the stat increment makes
future refactoring easier.

Change-Id: I87131c5517a4476c9a9b84263df107eca4686c60
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/191824
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>