MB-57104: Let SO_SNDBUF max size be tunable
To allow for increasing it on windows add the configuration
variable max_so_sndbuf_size in memcached.json.

The value defaults to 1MB for windows and 256MB (the old default)
for other platforms. The motivation for using different tunables
is to let people on the other platforms just increase the max value
in one location (the operating system) and not have to do a second
modification to Couchbase in order to use it.

Change-Id: Ibbc47b7b9aa82b028c8a8e2d6638c5685a591cab
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/191632
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>