MB-57106: Replace MagmaScanResult::Status::Next with existing Success
When we added ById scan support to MagmaKVStore (MB-52134) we
introduced Status::Next in the MagmaScanResult enumeration, along with
the canonical Status::Success that is used in other KV structures in
the backfill/scan components (eg, CouchKVStore names it
COUCHSTORE_SUCCESS).

That way we ended up using MagmaScanResult::Status::Next in place of
MagmaScanResult::Status::Success with exactly the same semantic.
Status::Success currently unused.

Rather than removing Status::Success, this patch changes MagmaKVStore
to using it in place of Status::Next that is also removed.

Change-Id: I4c9a7252d05680cb8816f10b7d053aca318ce3e1
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/191736
Well-Formed: Restriction Checker
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>