MB-56853: Move the collection operation counters to a single location
This change moves the 3 (u64) collection operation counters from the
vbucket data structure into a bucket data structure.

The aim of this patch is to reduce the memory footprint of the vbucket
collection management which improves our bucket overheads when a node
has many vbuckets and many collections.

With this commit the object size is reduced from 80 to 56 bytes.

A follow-up change will add CoreLocal wrappers to the counters as they
are written from many threads on hot paths.

Change-Id: I718a150c9734e62524e49e6795c07ad332bbb321
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/191755
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>