MB-56853: Change Collection startSeqno to cb::uint48_t
The aim of this commit is to reduce the size of the
Collections ::VB::ManifestEntry object  to reduce the memory overhead
of the collection configuration.

The maximum size of a seqno is 48-bits, so change the collection
start seqno to use this type and move it in the class definition so
the canDeduplicate byte can move into the trailing byte.

The object size is reduced from 56 to 48 bytes.

Change-Id: I91a0bc6ae331742dc244e77ad0fd9dc49ed7b1dc
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/191757
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>