MB-31806: Move Collection maxTTL to VB::ManifestEntry
In preparation for modify maxTTL support, the maxTTL must now be owned
by the VB::ManifestEntry and not the bucket shared metadata (similar
to CanDeduplicate).

This is required to ensure that when the maxTTL changes value each VB
can correctly detect the change and generate a modify event. If the
value was shared, the first updated vbucket would change the value
and prevent other vbuckets from making the correct modify decision.

Change-Id: I5458be277b6964cee5954cc10be30d9d7650542d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/191910
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>