MB-31806: VB::Manifest support for modifying collection TTL
Utilising the modifyCollection support added for history modifications,
detect if a collection's ttl has changed and generate the modify event.

As the modify event is a complete collection entry, only a few changes
are needed to get the new TTL replicated.

This commit only updates the VB::Manifest code and is only testable from
unit tests. A change of TTL from setCollections (and the bucket Manifest
level) will be rejected.

Change-Id: I4c320e408a56717d03deda2f97cd999670c41268
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/191911
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>